Set all domain ids to uint8

+++ chains/evm/cli/bridge/cancel-proposal.go
@@ -20,7 +20,7 @@ var cancelProposalCmd = &cobra.Command{
 func BindCancelProposalFlags(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&Bridge, "bridge", "", "Bridge contract address")
 	cmd.Flags().StringVar(&DataHash, "data-hash", "", "Hash of proposal metadata")
-	cmd.Flags().Uint64Var(&DomainID, "domain", 0, "Domain ID of proposal to cancel")
+	cmd.Flags().Uint8Var(&DomainID, "domain", 0, "Domain ID of proposal to cancel")
 	cmd.Flags().Uint64Var(&DepositNonce, "deposit-nonce", 0, "Deposit nonce of proposal to cancel")
 	flags.MarkFlagsAsRequired(cmd, "bridge", "data-hash", "domain", "deposit-nonce")
 }
